# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "Api_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/ready": {
      "get": {
        "operationId": "Api_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/auth/telegram": {
      "post": {
        "operationId": "Api_login",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Validate Telegram initData and create a cookie session",
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/auth/dev": {
      "post": {
        "operationId": "Api_dev",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/auth/dev-users": {
      "get": {
        "operationId": "Api_devUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "Api_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "Api_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/profile": {
      "put": {
        "operationId": "Api_profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/settings": {
      "put": {
        "operationId": "Api_settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/users/me": {
      "delete": {
        "operationId": "Api_deleteMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/photos": {
      "post": {
        "operationId": "Api_upload",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/photos/order": {
      "put": {
        "operationId": "Api_order",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/photos/{id}": {
      "delete": {
        "operationId": "Api_removePhoto",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/discovery": {
      "get": {
        "operationId": "Api_discover",
        "parameters": [
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/swipes": {
      "post": {
        "operationId": "Api_swipe",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/likes": {
      "get": {
        "operationId": "Api_likes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/conversations": {
      "get": {
        "operationId": "Api_chats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/conversations/{id}/messages": {
      "get": {
        "operationId": "Api_messages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      },
      "post": {
        "operationId": "Api_send",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/conversations/{id}/read": {
      "post": {
        "operationId": "Api_read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/conversations/{id}": {
      "delete": {
        "operationId": "Api_unmatch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/blocks/{id}": {
      "post": {
        "operationId": "Api_block",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      },
      "delete": {
        "operationId": "Api_unblock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/blocks": {
      "get": {
        "operationId": "Api_blocks",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/reports": {
      "post": {
        "operationId": "Api_report",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/plans": {
      "get": {
        "operationId": "Api_plans",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/payments/invoice": {
      "post": {
        "operationId": "Api_invoice",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/payments": {
      "get": {
        "operationId": "Api_paymentHistory",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/telegram/webhook": {
      "post": {
        "operationId": "Api_webhook",
        "parameters": [
          {
            "name": "x-telegram-bot-api-secret-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/content/{key}": {
      "get": {
        "operationId": "Api_content",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/admin/overview": {
      "get": {
        "operationId": "Api_overview",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/admin/{kind}": {
      "get": {
        "operationId": "Api_adminList",
        "parameters": [
          {
            "name": "kind",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/admin/photo/{id}": {
      "get": {
        "operationId": "Api_adminPhoto",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    },
    "/api/v1/admin/{kind}/{id}": {
      "put": {
        "operationId": "Api_adminChange",
        "parameters": [
          {
            "name": "kind",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Vela"
        ]
      }
    }
  },
  "info": {
    "title": "Vela API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vela_session"
      }
    },
    "schemas": {}
  }
}